Add gradle plugin to flyway project
A gradle plugin based on the Flyway 2.x series
([original](https://github.com/ben-manes/gradle-flyway-plugin)).

This plugin is integrated into the Maven build using the
[gradle-maven-plugin](https://github.com/if6was9/gradle-maven-plugin)
to invoke the gradle build. The gradle wrapper is invoked so that the
maven packaging phase has the compile jars in the target folder. This
allows maven to pick up the jars in subsequent phases, like install.