Add support for Apache Phoenix (http://phoenix.apache.org/)
It has been set up as an EmbeddedDB, although the Phoenix tests
should be disabled by default unless run in the new 'PhoenixDBTest'
profile.

Notable changes to core classes:

* MetaDataTableImpl:

Due to Phoenix's peculiar syntax, the hardcoded statements in
addAppliedMigration() and updateChecksum() won't work. It now
attempts to load a .sql template and run that, if it exists.

* MigrationTestCase:

Added checks for applied migrations of type MigrationType.SCHEMA
and adjusted asserts accordingly. Also added a 'getBaseDir()'
method, and replaced direct calls to BASEDIR with this. Several of
the tests in MigrationTestCase can then be used in the Phoenix
tests by overriding that method.