Escape command-line arguments.
Following the instructions at
http://flywaydb.org/documentation/existing.html on Linux, the following
command fails with `ERROR: Invalid operation: version`:

    ./flyway baseline -Dflyway.baselineVersion=1 \
        -Dflyway.baselineDescription="Base version"

The problem is that the `flyway` Bash script does not properly escape
the command line arguments.

This patch adds quotes around all variables to fix this error. As a
side effect, it also allows to have spaces in the Flyway installation
directory name.