Fixed issue with upgrading metadata table in AWS Redshift #1231 (#1256)
* Fixed #1231 upgrading AWS Redshift DB from Flyway 3.2.1 to 4.0 failed

Flyway couldn't upgrade the metadata table, due to the following error:
"Amazon Invalid operation: cannot insert/update into table after dropping
non-nullable column"

This was caused by trying to remove the NOT NULL constraint on the version
column by adding/dropping a temp columan and copying the values. For somae
unknown reason, this works on some versions of Redshift, but not others.
The solution is just to rename the original table, create a new one, copy
the data, and drop the original table.

* Configured SORTKEY for the metadata table in Redshift

Redshift does not support indexes because it is a column-oriented
database. Instead data can be physically stored sorted, to improve
reading and sorting performance, based on the sort key.

* Fixed the fix for #1231 upgrading AWS Redshift DB from Flyway 3.2.1 to 4.0 failed

When renaming a table, the schema should not be specified in the
new table name.