Updates Android scanning code to handle different dex locations and prevent crashes
Some phones store the dex files in different locations; we need to check two different directories at a minimum. We also need to load dex files differently to account for which directories are writeable. Finally, the dex check should not crash the app if we can't find those files; instead, just proceed without loading. It's not ideal but it avoids a crash the caller can do nothing about.