Throw exception if no migration found
Added property `checkLocation` for FlywayConfiguration. If this property is set to false
(default), then flyway will only log a warning if no migration script is found at a location.
If set to true, flyway will throw an exception. Unit tests for this feature are also added.

This feature is helpful if users want their program to fail fast when no migrations are detected
at a given location.

Fixes #1496