Fixed #1627 Continue classpath scan on AccessControlException
Some environments (like the Google App Engine Jetty Dev Server) pull
in JAR's from the host system that are not allowed to be scanned at
runtime.  This causes an AccessControlException to be thrown by
java.security, which aborts the classpath scan completely and causes
Flyway to fail.  These JAR's will never have DB migrations in them,
so the failure is unnecessary.  Let's catch this exception and just
warn the user instead.