Check if DBA_FLASHBACK_ARCHIVE_TABLES is granted
In disableFlashbackArchiveForFbaTrackedTables() we use DBA_FLASHBACK_ARCHIVE_TABLES, which is granted to PUBLIC by default. If it is not, there may be a error even if FBA is not used.

To avoid this, USER_ view now used if DBA_ view is not accessible and the cleaned schema is current user's own schema.