Add ability to ignore exceptions
Sometimes it is ok if a statement fails and we don't want the
entire migration to fail when this happens.

In Oracle SQL*Plus this is accomplished by surrounding blocks of statements with
```
WHENEVER SQLERROR CONTINUE;
[code where an exception should be printed but execution should continue]
WHENEVER SQLERROR EXIT FAILURE;
[code where an exception should cause the migration to stop and fail]
```
Flyway now recognizes these `WHENEVER SQLERROR` statements and adjusts
its behavior when an exception is encountered.