Changed log level for the happy path in doMigrateGroup() from debug to info
The line I changed was logged the completion of a task at the debug level, but the line logging the start of the task was at the info level, so it seemed to be missing from the log when I read it. All other logging in the class except this one line is also at the info level.