fix(gradle): make the flyway task classpath simpler
The strategy to compose the classpath of the flyway task was to add the classes and resources directories of all sourceSets, and then add the compile, runtime, testCompile and testRuntime dependencies. So, unless I'm mistaken, this basically boils down to using the whole test runtime classpath, which is accessible right away from the gradle API.

In addition, the code causes a warning since gradle 4.0, which now uses one separate classes output directory per language in a source set, instead of a single classes directory per source set. Using the test runtime classpath also allows avoiding this warning.

fixes #1710