always check that the lock was released
Leaving the lock in place can cause problems when using flyway from multiple processes or across multiple machines. It is also very hard to debug. This PR checks that the lock is actually released and if not output an error log.

An example of how this might happen: https://github.com/flyway/flyway/issues/2895