Do not use MySQL named locks on Galera clusters
Since MariaDB MDEV-30473 calling GET_LOCK on a (Galera) cluster is no
longer allowed. Before this change the call was not replicated and
only executed locally.

FlyWay now triggers the error "This version of MariaDB doesn't yet
support 'GET_LOCK in cluster (WSREP_ON=ON)'"

This change checks for WSREP_ON=ON and disables the use of named locks.