fix: only include user constraints
Filter on the table schema name when selecting all foreign key
constraints, to make sure that no system constraints are included in the
results if they were to be added in the future. This also makes sure
that if Cloud Spanner starts supporting named user schemas, the listed
foreign keys are consistent with the tables that are used, which are all
filtered based on an empty schema name.