Avoid generating FlywaySqlScriptException dynamically
This is suboptimal as we end up generating the same string over and over
again and this particular getMessage() requires class loader access.

In Quarkus, we actually call getMessage() after the class loader is
closed in some cases which causes some issues.

I think this is good practice anyway as better avoid calling the class
loader multiple times when an error occurs.

See https://github.com/quarkusio/quarkus/issues/43926 for the background
of this change.